Had to place the function better.
